move visualize and tests to new file
